Average now uses a sliding window for sample values (https://issues.redhat.com/browse/JGRP-2832)

+++ src/org/jgroups/util/Profiler.java
@@ -39,7 +39,7 @@ public class Profiler {
         Thread curr_thread=Thread.currentThread();
         Long start=threads.remove(curr_thread);
         if(start != null) {
-            long time=System.nanoTime() - start;
+            double time=System.nanoTime() - start;
             synchronized(avg) {
                 avg.add(time);
             }
